vision_pipeline/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
grpcio>=1.56.0
grpcio-tools>=1.56.0
protobuf>=4.23.0
pytest>=7.4.0
